services/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.6.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
httpx==0.25.2
openai==1.3.7
python-dotenv==1.0.0
email-validator==2.1.0
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1